+++ Marlin/Utility.h
@@ -24,7 +24,8 @@ char *ftostr32a(const float &x);
 char *ftostr(float x,int pre,int post, bool sign=false);
 char *itostr5(const int &xx);
 
-void SetLEDColor (int r, int g, int b, bool fade=true);
+void SetLEDColor (byte r, byte g, byte b, bool fade=true);
+
 unsigned long CalculateRemainingTime (float percent_complete,unsigned long elapsed_time);
 char* EchoTimeSpan (long t, bool shortform=false);
